+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Eliminate need for passing `area_draw_params` to `path` when declared underneath `area` `on_draw` listener
 - Support `path` `transform` property for building `matrix` objects (either through `path { transform {operations} }` directly or through `m1 = matrix {operations}; path { transform m1 }` to reuse a matrix)
 - Support `area` `transform` property for building `matrix` objects (either through `area { transform {operations} }` directly or through `m1 = matrix {operations}; area { transform m1 }` to reuse a matrix)
+- Ensure `matrix` `rotate` method accepts degrees (not radians)
 
 ## 0.1.4